requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
future==0.18.3
gevent==23.7.0
greenlet==2.0.2
idna==3.4
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==3.4.4
tzdata==2023.3
urllib3==2.0.4
zope.event==5.0
zope.interface==6.0
